Update index.html
Edits to the line: "Serving this group of remarkable Latter-day Saints is a truly joyful experience and often goes amazingly, even miraculously, smoothly; but not always 🙂. Sometimes it comes with inconvenience and frustration, as did Christ's own earthly ministry. For a little perspective, take a moment to read the volunteer driving anecdote and the short <a href="#stories-section">Stories of Jesus</a> below. You may find that they resonate and help you smile as you engage in this work."