Update index.html
Minor changes:

The Treasure Valley French-Swahili Branch is a vibrant and growing community of members of the Church of Jesus Christ of Latter-day Saints from mainly French and Swahili-speaking countries in Africa. Our members bring rich cultures, languages, experiences and a joyous love of the Gospel of Jesus Christ that strengthen our congregation.